pyproject.toml
[project]
name = "livekit-museum-curator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[deepgram,elevenlabs,openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
]
